scraper/requirements.txt
certifi==2019.11.28
chardet==3.0.4
Deprecated==1.2.7
idna==2.9
PyGithub==1.46
PyJWT==1.7.1
pymongo==3.10.1
requests==2.23.0
urllib3==1.25.8
wrapt==1.12.0
